Update README.md
I think your framework is awesome and I'd love to be apart of it if that's cool? I had this little playground logo I built in sketch awhile back laying around so I modified it and added the name. If you don't like it no worries! If you do, I'm more than happy to provide it! :)